+++ Krowi_AchievementFilter/Libs/KrowiTutorials-1.0/KrowiTutorials-1.0.lua
@@ -194,6 +194,7 @@ local function UpdateFrame(frame, i)
 			local image = frame:CreateTexture(); -- Implicitly calls Show
 			image:SetPoint("TOPLEFT", frame, imageMargin.Left, -imageMargin.Top);
 			image:SetWidth(imageSize.Width);
+			image:SetHeight(imageSize.Height);
 			image:SetTexture(page.Image);
 			image:SetTexCoord(imageTexCoord[1], imageTexCoord[2], imageTexCoord[3], imageTexCoord[4]);
 			page.ImageTex = image;
@@ -201,32 +202,33 @@ local function UpdateFrame(frame, i)
 	end
 	local fullImageWidth = page.Image and imageMargin.Left + imageSize.Width + imageMargin.Right or 0;
 	local fullImageHeight = page.Image and imageMargin.Top + imageSize.Height + imageMargin.Bottom or 0;
-	print("fullImageWidth " .. tostring(fullImageWidth))
+	-- print("fullImageWidth " .. tostring(fullImageWidth))
 
 	-- Text
 	local text = page.Text;
 	if page.SubTitle then
 		text = page.SubTitle .. "\n\n" .. text;
 	end
-	local textRelativePoint = page.Image and (layout == "HORIZONTAL" and "TOPRIGHT" or "BOTTOMLEFT") or "TOPLEFT";
 	local textRelativeFrame = page.Image and page.ImageTex or frame;
+	local textRelativePoint = page.Image and (layout == "HORIZONTAL" and "TOPRIGHT" or "BOTTOMLEFT") or "TOPLEFT";
 	local textOffsetX = (layout == "HORIZONTAL" and imageMargin.Right - default.ImageMargin.Right or 0) + textMargin.Left;
 	local textOffsetY = textMargin.Top + (page.Image and imageMargin.Bottom - default.ImageMargin.Bottom or 0);
+	local textWidth = textSize.Width ~= 0 and textSize.Width or max(imageSize.Width, width);
 	frame.Text:SetPoint("TOPLEFT", textRelativeFrame, textRelativePoint, textOffsetX, -textOffsetY);
-	frame.Text:SetWidth(layout == "HORIZONTAL" and textSize.Width or width - textMargin.Left - textMargin.Right);
+	frame.Text:SetWidth(layout == "HORIZONTAL" and textSize.Width or textWidth - textMargin.Left - textMargin.Right);
 	-- print(textSize.Width)
 	-- print(imageSize.Width)
 	-- print(width - textMargin.Left - textMargin.Right)
-	print("textSize.Width " .. tostring(textSize.Width))
-	print("imageSize.Width " .. tostring(imageSize.Width))
-	print("width " .. tostring(width))
-	print("textMargin.Left " .. tostring(textMargin.Left))
-	print("textMargin.Right " .. tostring(textMargin.Right))
+	-- print("textSize.Width " .. tostring(textSize.Width))
+	-- print("imageSize.Width " .. tostring(imageSize.Width))
+	-- print("width " .. tostring(width))
+	-- print("textMargin.Left " .. tostring(textMargin.Left))
+	-- print("textMargin.Right " .. tostring(textMargin.Right))
 	frame.Text:SetText(text);
 	local fullTextWidth = text and textMargin.Left + max(textSize.Width, frame.Text:GetWidth()) + textMargin.Right or 0;
 	fullTextWidth = fullTextWidth + (layout == "HORIZONTAL" and default.ImageMargin.Right or 0);
 	local fullTextHeight = text and textMargin.Top + max(textSize.Height, frame.Text:GetHeight()) + textMargin.Bottom or 0;
-	print("fullTextWidth " .. tostring(fullTextWidth))
+	-- print("fullTextWidth " .. tostring(fullTextWidth))
 
 	-- EditBox
 	-- if data.editbox then