+++ Krowi_AchievementFilter/Localization/GlobalStrings.lua
@@ -44,4 +44,9 @@ function globalStrings.Load(L)
     L["Version"] = GAME_VERSION_LABEL;
     L["Icon"] = EMBLEM_SYMBOL;
     L["Key Binding"] = KEY_BINDING;
+
+    -- Tutorial
+    L["Enabled"] = PVP_WAR_MODE_ENABLED;
+    L["Disabled"] = ADDON_DISABLED;
+    L["Show All %d Results"] = ENCOUNTER_JOURNAL_SHOW_SEARCH_RESULTS;
 end